Первый релиз
- Добавлен новый модуль, позволяющий получать список новостных статей с News API.
- В приложении теперь отображаются реальные новости.
- Реализовано многопоточное получение новостей от News API.
- Кнопка "Обновить" стала функционировать.
- Небольшая реорганизация проекта, распределение кода по нескольким файлам.
