+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "news_observer"
-version = "0.4.0"
+version = "1.0.0"
 edition = "2021"
 authors = ["Evgeny Mangasaryan <mangasaryan.ep@gmail.com>"]
 
@@ -8,9 +8,10 @@ authors = ["Evgeny Mangasaryan <mangasaryan.ep@gmail.com>"]
 eframe = "0.17.0"
 serde = { version = "1.0.137", features = ["derive"] }
 confy = "0.4.0"
-reqwest = "0.11.10"
+ureq = "2.4.0"
 tracing = "0.1.34"
 tracing-subscriber = "0.3.11"
+serde_json = "1.0.81"
 
 [profile.release]
 opt-level = 2
